+++ client/src/images/index.js
@@ -1,0 +1,6 @@
+import pingpong from "./pingpong.png";
+import tennis from "./tennis.png";
+import badminton from "./badminton.png";
+import main from "./main.png";
+
+export { pingpong, tennis, badminton, main };
