+++ client/src/styles/components/Post/_ImageCarousel.scss
@@ -11,10 +11,13 @@
 
     img {
       width: 100%;
+      height: 100%;
       border-radius: 15px;
+      object-fit: cover;
     }
   }
   .btn {
+    cursor: pointer;
     position: absolute;
     box-shadow: 0 0 2px $white;
     width: 50px;
