+++ client/src/styles/main.scss
@@ -26,8 +26,7 @@ $large: unquote("(min-width: 1600px)");
 @import "./components/Footer";
 
 // pages
-// @import "./pages/Landing";
-@import "./pages/Landing1";
+@import "./pages/Landing";
 @import "./pages/Login";
 @import "./pages/Signup";
 @import "./pages/Main";
