--- client/src/styles/pages/_Landing1.scss
@@ -1,140 +1,0 @@
-$landing-height: 600px;
-$content-width: 1000px;
-
-.landing--container {
-  width: 100%;
-
-  .btn {
-    @include button(50px, 0) {
-      width: 150px;
-      background-color: $sub-color;
-      color: white;
-      border-radius: 30px;
-      box-shadow: none;
-      margin-top: 20px;
-    }
-    span {
-      margin-top: 3px;
-    }
-  }
-
-  .landing--top {
-    @include flexbox();
-    width: 100%;
-    height: 700px;
-    background: $main-color;
-    background: radial-gradient(circle, $main-color 0%, #7cb379 100%);
-    position: relative;
-
-    .content--container {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      // border: 1px solid black;
-      width: 1000px;
-      .text--container {
-        display: flex;
-        flex-direction: column;
-      }
-      .title {
-        font-size: 2rem;
-        color: $white;
-        line-height: 45px;
-        font-weight: 600;
-      }
-      .text {
-        font-size: 1.2rem;
-        color: $white;
-        margin-top: 20px;
-      }
-    }
-
-    .img--container {
-      width: 300px;
-      height: 300px;
-      img {
-        width: 300px;
-        height: 300px;
-        object-fit: cover;
-      }
-    }
-
-    .wavedivider {
-      position: absolute;
-      bottom: 0;
-      left: 0;
-      width: 100%;
-      overflow: hidden;
-      line-height: 0;
-      transform: rotate(180deg);
-      svg {
-        position: relative;
-        display: block;
-        width: calc(103% + 1.3px);
-        height: 133px;
-        .shape-fill {
-          fill: #ffffff;
-        }
-      }
-    }
-  }
-
-  .landing--wrapper {
-    width: 100%;
-    // background-color: $white;
-    .container {
-      @include flexbox();
-      border-bottom: 1px solid black;
-      width: 100%;
-      height: $landing-height;
-
-      .content--container {
-        display: flex;
-        width: $content-width;
-        justify-content: space-between;
-        align-items: center;
-
-        .text--container {
-          display: flex;
-          flex-direction: column;
-          .title {
-            font-size: 1.8rem;
-            font-weight: 700;
-          }
-          .text {
-            font-size: 1.2rem;
-            margin-top: 20px;
-            line-height: 30px;
-          }
-        }
-        .img--container {
-          img {
-            width: 300px;
-            height: 300px;
-            object-fit: cover;
-          }
-        }
-      }
-
-      &:nth-child(2n) {
-        background-color: $white;
-        .content--container {
-          flex-direction: row-reverse;
-        }
-      }
-
-      &:last-child {
-        justify-content: center;
-        flex-direction: column;
-        height: 400px;
-        .content--container {
-          @include flexbox();
-          font-size: 1.5rem;
-        }
-        .btn--container {
-          margin-top: 40px;
-        }
-      }
-    }
-  }
-}
